Fixed a major for loop flaw in nn-descent and fixed data formating for new data

+++ nn_descentGo/benchmark.go
@@ -2,7 +2,7 @@ package main
 
 // this function is used to benchmark the accuracy based on the graph it recieves from main
 // this could also easily be multi-threaded but it would need structural changes in main "Rasmus" I might add this if necessary
-func benchmark(graph Graph) float64 {
+func benchmark(graph Graph) float32 {
 	totalCorrect := 0
 	totalPossible := graph.N * graph.K
 
@@ -20,13 +20,13 @@ func benchmark(graph Graph) float64 {
 		}
 	}
 
-	accuracy := float64(totalCorrect) / float64(totalPossible) * 100
+	accuracy := float32(totalCorrect) / float32(totalPossible) * 100
 	return accuracy
 }
 
 func getTrueNN(vertex int) []int {
 	truenn := make([]int, 0)
-	distanceList := make([]float64, 0)
+	distanceList := make([]float32, 0)
 	for i := 0; i < graph.N; i++ {
 		if i != vertex {
 			calculatedDistance := euclideanDistance(graph.Data[vertex*graph.Dim:(vertex+1)*graph.Dim], graph.Data[i*graph.Dim:(i+1)*graph.Dim])
@@ -47,11 +47,11 @@ func getTrueNN(vertex int) []int {
 }
 
 // returns in format [distance,Index]
-func findmax(slice []float64) []float64 {
-	max := []float64{0, 0} // [placement, distance]
+func findmax(slice []float32) []float32 {
+	max := []float32{0, 0} // [placement, distance]
 	for i, v := range slice {
 		if v > max[0] {
-			max[1] = float64(i)
+			max[1] = float32(i)
 			max[0] = v
 		}
 	}
